pyproject.toml
[project]
name = "clean-Fastapi"
version = "0.1.0"
author = [
    {name = "soyeonlee", email = "soyeonlee@publicai.co.kr"}
]
description = "FastAPI로 배우는 백엔드 프로그래밍 with 클린 아키텍쳐 : 모던 파이썬 웹 프레임워크 FastAPI로 TIL 앱 구축하기"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    # === FastAPI & Infra ===
    "fastapi[standard]==0.116.1",
    "passlib[argon2]==1.7.4",
    "dependency-injector==4.42.0",
    "asyncpg==0.30.0",
    "SQLAlchemy==2.0.36",
    "alembic==1.14.0",
    "pydantic-settings==2.6.1",
    "psycopg[binary]==3.2.3",
    "greenlet==3.1.1",
    "PyJWT[crypto]==2.9.0",
    # === Celery ===
    "celery==5.4.0",
    "kombu==5.3.7",
    "billiard==4.2.0",
    "amqp==5.2.0",
    "psycopg2-binary==2.9.9",
    # === 기타 ===
    "aiohttp>=3.8.0",
    "psutil>=7.0.0",
    "aio-pika>=9.5.7",
]

[project.optional-dependencies]
dev = [
    # === Test ===
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pytest-cov==6.0.0",
]

test = [
    # === Test ===
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pytest-cov==6.0.0",
]

lint = [
    # === Linting ===
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]

format = [
    # === Formatting ===
    "black>=24.0.0",
    "isort>=5.13.0",
]

all = [
    # === Test ===
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pytest-cov==6.0.0",
    # === Linting ===
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "mypy>=1.9.0",
]
